1.4.4.2: Initial compilation fixes

+++ patches/tModLoader/Terraria/ModLoader/ModSystem.cs
@@ -334,7 +334,7 @@ namespace Terraria.ModLoader
 		/// A more advanced option to PostWorldGen, this method allows you modify the list of Generation Passes before a new world begins to be generated. <para/>
 		/// For example, removing the "Planting Trees" pass will cause a world to generate without trees. Placing a new Generation Pass before the "Dungeon" pass will prevent the the mod's pass from cutting into the dungeon.
 		/// </summary>
-		public virtual void ModifyWorldGenTasks(List<GenPass> tasks, ref float totalWeight) { }
+		public virtual void ModifyWorldGenTasks(List<GenPass> tasks, ref double totalWeight) { }
 
 		/// <summary>
 		/// Use this method to place tiles in the world after world generation is complete.
