+++ patches/tModLoader/Terraria/ModLoader/SystemLoader.HookLists.cs
@@ -61,7 +61,7 @@ namespace Terraria.ModLoader
 
 		private delegate void DelegateModifyTimeRate(ref double timeRate, ref double tileUpdateRate, ref double eventUpdateRate);
 
-		private delegate void DelegateModifyWorldGenTasks(List<GenPass> passes, ref float totalWeight);
+		private delegate void DelegateModifyWorldGenTasks(List<GenPass> passes, ref double totalWeight);
 
 		private delegate bool DelegateHijackGetData(ref byte messageType, ref BinaryReader reader, int playerNumber);
